+++ src/test/java/com/prgrms/vouchermanagement/wallet/WalletRepositoryTest.java
@@ -13,18 +13,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.vouchermanagement.voucher.domain.Voucher;
 
 @SpringBootTest
-@ActiveProfiles("test")
 @Transactional
 public class WalletRepositoryTest {
 
 	@Autowired
-	WalletRepository walletRepository;
+	WalletJdbcRepository walletRepository;
 
 	private UUID existingCustomerId1;
 	private UUID existingCustomerId2;
